requirements.txt
langchain 
langchain-core 
langchain-community 
langchain-openai
langchain-groq
langchain-ollama
langchain-google-genai
langgraph
firecrawl-py 
unstructured
python-dotenv